Make up for missing upload RPC exceptions
**Phenomenon and reproduction steps**

Missing upload RPC exception

**Root cause and solution**

Make up for missing upload RPC exceptions

**Impactions**

none

**Test method**

pass all unit test

**Affected branch(es)**

* main

**Checklist**

- [ ] Change synchronize required (trident-dump, trident-copy or trident-testsrv)
- [ ] Dependencies update required
- [ ] Common bug (similar problem in other code)
- [ ] Package successful (has changes in Makefile)

+++ agent/src/rpc/mod.rs
@@ -11,7 +11,7 @@ pub fn get_timestamp(ntp_diff: i64) -> Duration {
     let now = SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap()
-        .as_nanos() as u64 as i64
-        + ntp_diff;
+        .as_nanos() as i128
+        + ntp_diff as i128;
     Duration::from_nanos(now as u64)
 }
